Update Enigmail keyserver settings
Assume that a user has dirmngr running and configured with `use-tor' in
dirmngr.conf and set the relevant keyserver settings (similar to what
Tails has been doing); otherwise the keyserver lookup will fail-safe.
This should work for the daemon and Tor Browser since as per
gnupg/dirmngr/dns-stuff.c, it tries both 9050 and 9150. This is probably
better than the older setting because a) it actually works now, b) it
will still fail-safe in case Tor is not running.